Guard against indefinite caching
Because the timeout value for the ZGW caching is taken from the
Django settings, it is not unthinkable that, due to some
misconfiguration, the timeout value could end up being None. If
this is passed as-is to the low-level Django caching backend, it
will be interpreted as "cache forever", because the default kwarg
of 60 would be overridden. This can lead to unexpected behavior,
e.g. ZGW API objects appearing to be stale because the cache
never expires.

This commit makes sure an exception is raised if a non-integer
value is provided.